requirements.txt
pytest
selenium
opencv-python
numpy
Pillow